fix: remove -Wno-sign-compare and fix resulting warnings
Remove the -Wno-sign-compare suppression flag so sign-compare
warnings are caught by -Wall -Werror in DEBUG builds.

Fix all resulting warnings:
- ext.cpp: use size_t for loop iterators comparing against
  FThetaPolynomialDegree (which is size_t)
- CameraWrappers.cu: use static_cast<int> for comparing
  size_t index against int difference in normalize_shape()

+++ gsplat/cuda/csrc/CameraWrappers.cu
@@ -54,7 +54,7 @@ torch::Tensor normalize_shape(torch::Tensor tensor, std::string_view tensor_name
         // Reshape by prepending 1's if too few dimensions
         if(tensor.dim() < expected_ndims)
         {
-            tensor = tensor.expand({((Is < expected_ndims - tensor.dim()) ? 1 : -1)...});
+            tensor = tensor.expand({((static_cast<int>(Is) < expected_ndims - static_cast<int>(tensor.dim())) ? 1 : -1)...});
         }
 
         // Flatten leftmost dimensions if too many, or use as-is
